V2 (With Image)/App/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
faiss-cpu==1.9.0
filelock==3.13.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightning-utilities==0.11.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mlxtend==0.23.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.12.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyDeprecate==0.3.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rouge-score==0.1.2
rpds-py==0.20.1
rsa==4.9
sacremoses==0.1.1
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
simplet5==0.1.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.39.0
streamlit-option-menu==0.4.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchmetrics==1.5.1
torchvision==0.20.1+cu118
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.16.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.1.0
yarl==1.17.1